requirements.txt
Flask==2.0.1
Flask-SocketIO==5.2.0
eventlet==0.33.0
transformers==4.25.1
torch
celery==5.2.7
redis==4.3.4
SQLAlchemy==1.4.39
requests==2.28.1
Pillow==9.2.0
selenium==4.8.0
psutil==5.9.4
pytest==7.3.1
flake8==5.0.4
autopep8==1.7.0
reportlab==3.6.12
python-docx==0.8.11
openpyxl==3.1.2
moviepy==1.0.3
